Centralize the calendar conversion to/from byte array take 2.
Move to an independent class.
